requirements.txt
flask==3.0.0
pandas==2.1.4
python-docx==1.1.0
openpyxl==3.1.2
pyinstaller==6.3.0
